refactor gallery
